rename crypto store
